requirements.txt
requests==2.25.1
discord.py==1.6.0